app/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
ndg-httpsclient==0.5.1
oauthlib==3.2.1
packaging==21.3
pusher==3.2.0
pyasn1==0.4.8
pycparser==2.21
psycopg2-binary==2.9.5
PyJWT==1.7.1
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
pillow==9.2.0
boto3